Fixed code review comments
